made accountCreationForm a bit more lean
